# openapi.yaml
openapi: 3.1.0
info:
  title: FluxTrackr API
  version: 1.0.0
  description: |
    Contrato HTTP da API NestJS do FluxTrackr. A especificação foi derivada dos
    controllers, DTOs, serviços e schema Prisma em 2026-07-16. Não há prefixo
    global de rota: os caminhos abaixo são relativos diretamente à origem.

    Todas as datas e horas usam ISO 8601. Valores retornados pelo domínio
    financeiro são strings decimais com duas casas quando o cálculo precisa
    preservar precisão; os DTOs legados que recebem `number` continuam assim.
servers:
  - url: http://localhost:3000
    description: Desenvolvimento padrão
  - url: http://localhost:3001
    description: Desenvolvimento com PORT=3001
...
security:
  - bearerAuth: []
paths:
...
  /me:
    get:
      tags: [Me]
      operationId: getMe
      summary: Dados do usuário autenticado
      responses:
        '200': { description: Usuário autenticado, content: { application/json: { schema: { $ref: '#/components/schemas/Me' } } } }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Me]
      operationId: updateMe
      summary: Atualiza o nome do usuário autenticado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeRequest'
      responses:
        '200': { description: Usuário atualizado, content: { application/json: { schema: { $ref: '#/components/schemas/Me' } } } }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /me/password:
    patch:
      tags: [Me]
      operationId: changePassword
      summary: Troca a senha do usuário autenticado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200': { description: Senha atualizada, content: { application/json: { schema: { $ref: '#/components/schemas/PasswordUpdated' } } } }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token retornado por `POST /auth/login`.
...
  responses:
    BadRequest:
      description: Payload, parâmetro, regra de negócio ou cursor inválido
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: JWT ausente, inválido ou credenciais inválidas
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    InternalServerError:
      description: Erro inesperado do servidor
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    Uuid:
      type: string
      format: uuid
    DateTime:
      type: string
      format: date-time
...
    ErrorResponse:
      type: object
      required: [statusCode, message]
      properties:
        statusCode: { type: integer, examples: [400] }
        message:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
        error: { type: string }
...
    Me:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        name: { type: string }
        email: { type: string, format: email }
        createdAt: { $ref: '#/components/schemas/DateTime' }
    UpdateMeRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, minLength: 1 }
        newPassword: { type: string, minLength: 8, maxLength: 128 }
    PasswordUpdated:
      type: object
      required: [updated]
      properties:
        updated: { type: boolean, enum: [true] }
...